code style fixes
